Introduce a `should-fail` and a `should-pass` test case
