+++ tests/should-fail/package.json
