+++ tests/should-fail/src/vite-env.d.ts
