+++ tests/should-fail/vite.config.ts
