+++ tests/should-pass/eslint.config.js
@@ -1,0 +1,8 @@
+import config from "../../lib/react.js";
+
+export default [
+    ...config,
+    {
+        ignores: ["eslint.config.js"],
+    },
+];
