+++ tests/should-pass/package.json
@@ -1,0 +1,22 @@
+{
+    "dependencies": {
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1"
+    },
+    "devDependencies": {
+        "@types/react": "^18.3.12",
+        "@types/react-dom": "^18.3.1",
+        "@vitejs/plugin-react": "^4.3.3",
+        "vite": "^5.4.10"
+    },
+    "name": "@prezly/code-style-test-react-ts",
+    "private": true,
+    "scripts": {
+        "build": "tsc -b && vite build",
+        "dev": "vite",
+        "lint": "eslint .",
+        "preview": "vite preview"
+    },
+    "type": "module",
+    "version": "0.0.0"
+}
