+++ tests/should-pass/vite.config.ts
@@ -1,0 +1,8 @@
+import react from "@vitejs/plugin-react";
+import { defineConfig } from "vite";
+
+// https://vite.dev/config/
+// eslint-disable-next-line import/no-default-export
+export default defineConfig({
+    plugins: [react()],
+});
